==> scripts/07-bom-stations.R <==
library(tidyverse)
wind_raw <- tribble(
  ~Site,    ~Name,                                     ~Lat,       ~Lon,      ~Start,    ~End,       ~Years,   ~percent,      ~Obs, ~AWS,

  85000, "ABERFELDY",                                -37.7000,  146.3667, "Sep 1969", "Oct 1974",    5.2,   95,     1.9,   "N",
  90180, "AIREYS INLET",                             -38.4583,  144.0883, "Jul 1990", "Nov 2021",   31.4,   99,    12.1,   "Y",
  88001, "ALEXANDRA (POST OFFICE)",                  -37.1916,  145.7116, "Jan 1965", "Feb 1970",    5.2,   83,     1.0,   "N",
  89000, "ARARAT POST OFFICE",                       -37.2833,  142.9500, "Jan 1962", "Apr 1969",    7.3,   97,     2.0,   "N",
  89085, "ARARAT PRISON",                            -37.2769,  142.9786, "Jun 1969", "Nov 2021",   52.5,  100,     3.9,   "N",
  86147, "ASPENDALE CSIRO",                          -38.0333,  145.1000, "Jan 1965", "Jun 1982",   17.2,   85,     1.5,   "N",
  87113, "AVALON AIRPORT",                           -38.0288,  144.4783, "Nov 1965", "Nov 2021",   31.7,   55,    11.4,   "Y",
  81000, "AVOCA (POST OFFICE)",                      -37.0887,  143.4747, "Jan 1965", "Aug 1975",    9.6,   70,     1.0,   "N",
  84108, "BAIRNSDALE",                               -37.8333,  147.6500, "Apr 1970", "Jun 1983",   12.6,   92,     2.0,   "N",
  85279, "BAIRNSDALE AIRPORT",                       -37.8818,  147.5669, "Jul 1942", "Nov 2021",   41.7,   52,     8.4,   "Y",
  84080, "BAIRNSDALE POST OFFICE",                   -37.8167,  147.6167, "Jan 1957", "Mar 1970",   13.3,   99,     2.0,   "N",
  87005, "BALLAN (FISKVILLE)",                       -37.6000,  144.2000, "Jan 1957", "Jun 1969",   12.5,   96,     2.0,   "N",
  89002, "BALLARAT AERODROME",                       -37.5127,  143.7911, "Jul 1951", "Nov 2021",   61.8,   87,     7.1,   "Y",
  82001, "BEECHWORTH COMPOSITE",                     -36.3702,  146.7132, "Jan 1957", "Nov 1986",   29.9,   97,     1.9,   "N",
  82137, "BEECHWORTH WOOLSHED",                      -36.3192,  146.6803, "Nov 1986", "Apr 2012",   13.2,   49,     2.0,   "N",
  82002, "BENALLA (SHADFORTH STREET)",               -36.5483,  145.9703, "Jan 1957", "Oct 2006",   46.5,   92,     3.6,   "N",
  82170, "BENALLA AIRPORT",                          -36.5522,  145.9977, "Nov 2006", "Nov 2021",   15.1,   98,     2.6,   "N",
  81123, "BENDIGO AIRPORT",                          -36.7411,  144.3274, "Oct 1991", "Nov 2021",   30.2,   99,    11.5,   "Y",
  81003, "BENDIGO PRISON",                           -36.7533,  144.2825, "Jan 1957", "Oct 1991",   34.8,   99,     2.6,   "N",
  84077, "BENDOC FORESTRY",                          -37.1500,  148.8833, "Jan 1965", "Sep 1967",    2.6,   61,     1.0,   "N",
  77004, "BEULAH",                                   -35.9384,  142.4202, "Jan 1957", "Sep 1975",   18.8,   98,     2.0,   "N",
  77007, "BIRCHIP POST OFFICE",                      -35.9825,  142.9156, "Jan 1965", "Jan 1984",   19.0,   79,     1.4,   "N",
  87159, "BLACKWOOD (ARMSTRONG)",                    -37.4667,  144.3167, "Dec 1974", "Dec 1982",    8.0,   98,     2.0,   "N",
  83004, "BOGONG",                                   -36.8056,  147.2239, "Jan 1962", "Sep 1975",   13.8,   90,     2.0,   "N",
  82100, "BONEGILLA",                                -36.1333,  147.0000, "Jun 1968", "Jun 1986",   18.1,  100,     2.0,   "N",
  80002, "BOORT",                                    -36.1160,  143.7229, "Jan 1965", "Nov 1986",   21.7,   82,     1.7,   "N",
  87184, "BREAKWATER (GEELONG RACECOURSE)",          -38.1737,  144.3765, "Jun 2011", "Nov 2021",   10.5,  100,    20.6,   "Y",
  83067, "BRIGHT",                                   -36.7313,  146.9614, "Jun 1969", "Dec 1994",   25.6,   95,     2.0,   "N",
  86033, "BRIGHTON BOWLS CLUB",                      -37.9252,  145.0254, "Jan 1965", "Nov 1967",    2.9,   98,     1.1,   "N",
  86351, "BUNDOORA (LATROBE UNIVERSITY)",            -37.7163,  145.0453, "Sep 1979", "Dec 2018",   39.3,   93,     1.2,   "N",
  90011, "CAMPERDOWN (POST OFFICE)",                 -38.2331,  143.1450, "Jan 1965", "Aug 1975",   10.7,   81,     1.0,   "N",
  84027, "CANN RIVER FORESTRY",                      -37.5333,  149.1500, "Jan 1965", "Dec 1973",    7.0,   56,     1.8,   "N",
  90184, "CAPE NELSON LIGHTHOUSE",                   -38.4306,  141.5437, "Dec 1995", "Nov 2021",   26.0,   99,    12.5,   "Y",
  90014, "CAPE NELSON LIGHTHOUSE COMPARISON",        -38.4319,  141.5419, "Jan 1957", "Jun 1999",   42.5,   99,     2.2,   "N",
  90015, "CAPE OTWAY LIGHTHOUSE",                    -38.8557,  143.5130, "Jan 1861", "Nov 2021",  160.8,   99,     3.7,   "Y",
  86017, "CAPE SCHANCK LIGHTHOUSE",                  -38.4931,  144.8858, "Jan 1962", "Aug 1997",   35.5,   89,     1.5,   "N",
  86245, "CARRUM DOWNS",                             -38.1167,  145.1797, "Dec 1967", "Oct 1975",    7.9,   67,     1.9,   "N",
  90182, "CASTERTON",                                -37.5831,  141.3340, "Mar 2005", "Nov 2021",   16.8,   99,    12.8,   "Y",
  90135, "CASTERTON SHOWGROUNDS",                    -37.5909,  141.4124, "Jan 1957", "Feb 2011",   54.2,   98,     1.9,   "N",
  88014, "CASTLEMAINE POST OFFICE",                  -37.1000,  144.2000, "Dec 1965", "Feb 1966",    0.3,   34,     1.0,   "N",
  88110, "CASTLEMAINE PRISON",                       -37.0809,  144.2391, "Mar 1966", "Nov 2021",   55.8,   98,     1.9,   "N",
  86361, "CERBERUS",                                 -38.3646,  145.1785, "Oct 1987", "Nov 2021",   34.2,   99,    11.6,   "Y",
  80128, "CHARLTON",                                 -36.2846,  143.3341, "Jul 2004", "Nov 2021",   17.4,  100,    14.9,   "Y",
  80006, "CHARLTON POST OFFICE",                     -36.3000,  143.4000, "Jan 1965", "Sep 1971",    6.8,   95,     1.9,   "N",
  88015, "CLUNES",                                   -37.3049,  143.7776, "Jan 1965", "Sep 1975",   10.8,   81,     1.0,   "N",
  90174, "COLAC (ELLIMINYT)",                        -38.3933,  143.5961, "May 1983", "Oct 1998",   15.5,   98,     2.7,   "N",
  90035, "COLAC (MOUNT GELLIBRAND)",                 -38.2332,  143.7924, "Jul 2000", "Nov 2021",   21.4,   99,    14.2,   "Y",
  90147, "COLAC (SHIRE OFFICE)",                     -38.3427,  143.5848, "Jan 1957", "Apr 1983",   23.8,   89,     2.0,   "N",
  90022, "COLAC AIRPORT",                            -38.2798,  143.6613, "Oct 1998", "Aug 2000",    1.9,   98,     3.0,   "N",
  86383, "COLDSTREAM",                               -37.7239,  145.4092, "Oct 1994", "Nov 2021",   27.2,   99,    12.3,   "Y",
  84143, "COMBIENBAR AWS",                           -37.3416,  149.0227, "Jul 1994", "Nov 2021",   27.3,   98,    12.7,   "Y",
  85301, "CORNER INLET (YANAKIE)",                   -38.8051,  146.1936, "Jan 2013", "Nov 2021",    8.9,   99,    22.8,   "Y",
  82011, "CORRYONG (PARISH LANE)",                   -36.2003,  147.8956, "Feb 1972", "Mar 2006",   34.2,   99,     3.0,   "N",
  82169, "CORRYONG AIRPORT",                         -36.1826,  147.8942, "Mar 2006", "Nov 2021",   15.6,   94,     1.6,   "N",
  82066, "CORRYONG FORESTRY",                        -36.2000,  147.9000, "Jan 1965", "Jan 1972",    6.5,   69,     1.8,   "N",
  86375, "CRANBOURNE BOTANIC GARDENS",               -38.1348,  145.2637, "Feb 1990", "Mar 2020",   29.6,   93,     1.0,   "N",
  88019, "CRESWICK",                                 -37.4200,  143.8924, "Jan 1965", "Sep 1975",   10.7,   96,     1.0,   "N",
  86368, "DALMORE",                                  -38.1800,  145.4317, "Jun 1988", "Aug 1990",    2.3,   83,     1.6,   "N",
  86224, "DANDENONG",                                -37.9764,  145.2247, "Jan 1964", "Nov 1974",    9.6,   79,     1.8,   "N",
  85275, "DARRIMAN (OMEGA)",                         -38.4833,  146.9333, "Apr 1981", "Feb 1982",    0.9,   96,     2.0,   "N",
  90194, "DARTMOOR",                                 -37.9222,  141.2614, "Mar 2009", "Nov 2021",   12.8,  100,    18.3,   "Y",
  82076, "DARTMOUTH RESERVOIR",                      -36.5352,  147.4984, "Jan 1975", "Nov 2021",   46.8,   94,     1.0,   "N",
  78072, "DONALD",                                   -36.3611,  143.0038, "Oct 1966", "Nov 2000",   34.2,   98,     2.0,   "N",
  81013, "DOOKIE AGRICULTURAL COLLEGE",              -36.3717,  145.7048, "Jan 1965", "Oct 1975",   10.8,   99,     1.0,   "N",
  87021, "DURDIDWARRAH",                             -37.8233,  144.2078, "Jan 1965", "Jul 2000",   35.6,  100,     1.0,   "N",
  85072, "EAST SALE",                                -38.1156,  147.1323, "Apr 1943", "Sep 2017",   74.5,   99,     7.9,   "Y",
  85314, "EAST SALE AIRPORT",                        -38.1016,  147.1398, "Oct 2017", "Nov 2021",    4.2,  100,    21.0,   "Y",
  85227, "EAST TARWIN (MIRBOO PASTORAL COMPANY)",    -38.5215,  146.2041, "Jan 1975", "Jun 1984",    9.5,   97,     2.0,   "N",
  85235, "EAST TARWIN NO.18",                        -38.5269,  146.2581, "May 1972", "Dec 1974",    2.6,   92,     2.0,   "N",
  80015, "ECHUCA AERODROME",                         -36.1648,  144.7643, "Jan 1960", "Nov 2021",   61.8,   98,     1.9,   "N",
  79099, "EDENHOPE AIRPORT",                         -37.0222,  141.2657, "Apr 2005", "Nov 2021",   16.7,   99,    15.9,   "Y",
  83083, "EDI UPPER",                                -36.7396,  146.4671, "Nov 1985", "Feb 2020",   33.8,   77,     1.6,   "N",
  88164, "EILDON FIRE TOWER",                        -37.2091,  145.8423, "Jul 1996", "Nov 2021",   25.4,   99,    13.2,   "Y",
  85240, "ELLINBANK DAIRY RESEARCH INST",            -38.2456,  145.9347, "Aug 1983", "Aug 2000",   17.1,   82,     1.6,   "N",
  85026, "ERICA",                                    -37.9747,  146.3745, "Jan 1965", "Sep 1975",   10.8,   95,     2.0,   "N",
  86038, "ESSENDON AIRPORT",                         -37.7276,  144.9066, "Apr 1939", "Nov 2021",   52.2,   63,     9.5,   "Y",
  82016, "EUROA",                                    -36.7569,  145.5660, "Jan 1965", "Feb 1976",   11.2,   95,     1.7,   "N",
  83084, "FALLS CREEK",                              -36.8708,  147.2755, "Sep 1990", "Nov 2021",   30.6,   94,    11.3,   "Y",
  83071, "FALLS CREEK",                              -36.8667,  147.2833, "Jul 1970", "Sep 1970",    0.3,   78,     1.8,   "N",
  83037, "FALLS CREEK SEC",                          -36.8633,  147.2794, "Apr 1974", "Sep 1990",    4.9,   28,     2.0,   "N",
  86376, "FAWKNER BEACON",                           -37.9483,  144.9269, "Mar 2000", "Nov 2021",   21.5,   97,    23.6,   "Y",
  86266, "FERNY CREEK",                              -37.8748,  145.3497, "Mar 2011", "Nov 2021",   10.8,   99,    20.3,   "Y",
  86372, "FERNY CREEK (DUNNS HILL)",                 -37.8775,  145.3364, "Nov 1990", "Jul 2011",   20.8,   98,     7.9,   "Y",
  90040, "FORREST STATE FOREST",                     -38.5198,  143.7150, "Jan 1965", "Aug 1975",   10.6,   68,     1.9,   "N",
  86388, "FRANKSTON (BALLAM PARK)",                  -38.1517,  145.1615, "Jun 2020", "Nov 2021",    1.5,   98,    23.7,   "Y",
  86371, "FRANKSTON BEACH",                          -38.1480,  145.1152, "Feb 1991", "Nov 2021",   30.2,   96,    12.3,   "Y",
  86240, "FRENCH ISLAND (MCLEOD PRISON)",            -38.4000,  145.3000, "Oct 1969", "Apr 1975",    5.6,   98,     2.0,   "N",
  84016, "GABO ISLAND LIGHTHOUSE",                   -37.5679,  149.9158, "Jan 1957", "Nov 2021",   64.9,   99,     5.3,   "Y",
  87117, "GEELONG (NORLANE)",                        -38.0850,  144.3356, "Jun 1970", "Jul 1996",   26.2,   97,     2.0,   "N",
  87080, "GEELONG (POINT HENRY)",                    -38.1333,  144.4167, "May 1984", "Mar 1985",    0.9,   97,     2.0,   "N",
  87023, "GEELONG SALINES (MOOLAP)",                 -38.1587,  144.3954, "Jan 1965", "Aug 1967",    2.7,   98,     1.0,   "N",
  87025, "GEELONG SEC",                              -38.1167,  144.3667, "Jan 1964", "Jun 1970",    6.5,   99,     2.0,   "N",
  84142, "GELANTIPY",                                -37.2200,  148.2626, "Dec 1992", "Nov 2021",   29.0,   98,    12.0,   "Y",
  84017, "GELANTIPY EAST",                           -37.2500,  148.3000, "Jul 1938", "May 1944",    5.8,   94,     1.0,   "N",
  90134, "GELLIBRAND RIVER FORESTRY",                -38.5304,  143.5390, "Jan 1965", "Oct 1981",   16.8,   91,     1.8,   "N",
  81044, "GOULBURN RIVER @ SHEPPARTON",              -36.3783,  145.4192, "Jan 1965", "Jun 1972",    7.5,   69,     1.0,   "N",
  79103, "GRAMPIANS (MOUNT WILLIAM)",                -37.2950,  142.6039, "Dec 2005", "Nov 2021",   15.8,   97,    16.1,   "Y",
  87163, "GROVEDALE (GEELONG AIRPORT)",              -38.2241,  144.3345, "Oct 1987", "Jun 2011",   23.8,   98,     7.7,   "Y",
  90044, "HAMILTON",                                 -37.7333,  142.0167, "Jan 1957", "Jun 1983",   24.9,   94,     2.0,   "N",
  90173, "HAMILTON AIRPORT",                         -37.6486,  142.0636, "Jul 1983", "Nov 2021",   38.1,   98,    10.7,   "Y",
  90169, "HAMILTON GODFREY",                         -37.7167,  141.9667, "Jul 1980", "Nov 1981",    1.4,   96,     2.0,   "N",
  90103, "HAMILTON RESEARCH STATION",                -37.8250,  142.0644, "Jan 1965", "May 2000",   35.1,   93,     1.0,   "N",
  86050, "HEALESVILLE (BADGER CRK SANCTUARY)",       -37.6833,  145.5333, "Jul 1973", "Sep 1990",    6.3,   30,     1.8,   "N",
  88029, "HEATHCOTE",                                -36.9589,  144.7010, "Aug 1969", "Sep 1975",    5.8,   81,     1.9,   "N",
  90048, "HEYWOOD FORESTRY",                         -38.1355,  141.6316, "Jan 1965", "Mar 2009",   39.3,   73,     1.9,   "N",
  77010, "HOPETOUN AIRPORT",                         -35.7151,  142.3569, "Jun 2004", "Nov 2021",   17.5,   99,    15.6,   "Y",
  79100, "HORSHAM AERODROME",                        -36.6699,  142.1733, "Oct 1997", "Nov 2021",   24.2,   99,    13.2,   "Y",
  79023, "HORSHAM POLKEMMET RD",                     -36.6613,  142.0687, "Jan 1957", "Apr 2012",   54.8,   98,     2.7,   "N",
  83014, "HOTHAM HEIGHTS",                           -36.9833,  147.1500, "Jun 1967", "Dec 1972",    3.1,   43,     1.9,   "N",
  82139, "HUNTERS HILL",                             -36.2137,  147.5395, "Apr 1993", "Nov 2021",   28.6,   98,    12.5,   "Y",
  78086, "JEPARIT",                                  -36.1442,  141.9928, "Jan 1965", "Sep 1975",   10.8,   87,     1.5,   "N",
  79097, "KANAGULK",                                 -37.1169,  141.8031, "May 2004", "Nov 2021",   17.6,   99,    14.8,   "Y",
  78018, "KELLALAC (WARRACKNABEAL AIRPORT)",         -36.3204,  142.4161, "Dec 2016", "Nov 2021",    5.0,   99,    23.9,   "Y",
  80023, "KERANG",                                   -35.7236,  143.9196, "Jan 1962", "Nov 2021",   56.8,   92,     2.0,   "N",
  86060, "KEW",                                      -37.8167,  145.0333, "Jan 1965", "Jun 1973",    8.0,   88,     1.0,   "N",
  80091, "KYABRAM",                                  -36.3350,  145.0638, "Jan 1965", "Nov 2021",   56.9,   95,     5.5,   "Y",
  88123, "KYNETON",                                  -37.2500,  144.4378, "Jul 1969", "Feb 1976",    6.7,   98,     2.0,   "N",
  88036, "KYNETON POST OFFICE",                      -37.2500,  144.4500, "Jan 1957", "May 1966",    9.4,   84,     1.9,   "N",
  88023, "LAKE EILDON",                              -37.2313,  145.9124, "Mar 1970", "Nov 2021",   51.8,   99,     1.9,   "N",
  84083, "LAKES ENTRANCE",                           -37.8692,  147.9961, "Sep 1965", "May 2006",   40.8,   96,     2.8,   "N",
  84150, "LAKES ENTRANCE (EASTERN BEACH ROAD)",      -37.8717,  148.0060, "Jun 2006", "Nov 2021",   15.5,   93,     2.2,   "N",
  87031, "LAVERTON RAAF",                            -37.8565,  144.7565, "Aug 1941", "Nov 2021",   75.8,   93,     8.6,   "Y",
  87032, "LAVERTON SALINES",                         -37.9000,  144.7500, "Jan 1965", "Apr 1965",    0.3,   66,     1.0,   "N",
  81084, "LEMNOS (CAMPBELLS SOUP)",                  -36.3608,  145.4594, "Jan 1965", "Jan 1996",   31.1,  100,     2.0,   "N",
  89018, "LISMORE (POST OFFICE)",                    -37.9543,  143.3437, "Jan 1957", "Mar 1994",   37.2,   90,     1.7,   "N",
  79028, "LONGERENONG",                              -36.6722,  142.2991, "Jan 1965", "Nov 2021",   56.9,   93,     6.7,   "Y",
  89105, "LOOKOUT HILL",                             -37.2822,  143.2469, "Dec 1991", "Apr 2007",   15.1,   94,     7.7,   "Y",
  90148, "LORNE PIER HEAD",                          -38.5333,  143.9833, "Dec 1969", "Apr 1986",   16.4,   96,     1.9,   "N",
  87036, "MACEDON FORESTRY",                         -37.4155,  144.5569, "Jul 1969", "Sep 1999",   30.0,   76,     1.8,   "N",
  85055, "MAFFRA FORESTRY OFFICE",                   -37.9664,  146.9831, "Jan 1965", "Aug 1975",   10.6,   76,     1.6,   "N",
  88041, "MALDON (DERBY HILL)",                      -36.9989,  144.0703, "Jan 1965", "Sep 1975",   10.6,   79,     1.0,   "N",
  84084, "MALLACOOTA",                               -37.5976,  149.7289, "Apr 1993", "Nov 2021",   28.6,   98,    11.0,   "Y",
  88111, "MALMSBURY",                                -37.2000,  144.4000, "Jul 1966", "Jun 1969",    3.0,   93,     1.9,   "N",
  88109, "MANGALORE AIRPORT",                        -36.8886,  145.1859, "Jan 1959", "Nov 2021",   62.6,   93,     7.2,   "Y",
  88043, "MARYBOROUGH",                              -37.0560,  143.7320, "Jan 1965", "Nov 2021",   56.8,   98,     1.8,   "N",
  86338, "MELBOURNE (OLYMPIC PARK)",                 -37.8255,  144.9816, "May 2013", "Nov 2021",    8.6,   99,    21.6,   "Y",
  86282, "MELBOURNE AIRPORT",                        -37.6654,  144.8322, "Jul 1970", "Nov 2021",   51.4,  100,    10.5,   "Y",
  86384, "MELBOURNE AIRPORT COMPARISON",             -37.6750,  144.8419, "Aug 1999", "Jan 2001",    1.5,   90,     5.7,   "Y",
  86071, "MELBOURNE REGIONAL OFFICE",                -37.8075,  144.9700, "Jan 1955", "Sep 2011",   54.8,   96,     8.0,   "Y",
  76026, "MERBEIN CSIRO RESEARCH STATION",           -34.2133,  142.0400, "Jan 1965", "Oct 1975",   10.8,  100,     1.0,   "N",
  76031, "MILDURA AIRPORT",                          -34.2358,  142.0867, "Jan 1946", "Nov 2021",   75.7,   99,     9.1,   "Y",
  76077, "MILDURA POST OFFICE",                      -34.1833,  142.2000, "Jun 1889", "Dec 1949",   59.9,   83,     1.3,   "N",
  82068, "MITTA MITTA",                              -36.5293,  147.3748, "Jan 1965", "Feb 1972",    7.2,   73,     1.0,   "N",
  85199, "MOONDARRA RESERVOIR",                      -38.0992,  146.3792, "Mar 1976", "Feb 1988",   12.0,   94,     1.0,   "N",
  86077, "MOORABBIN AIRPORT",                        -37.9800,  145.0962, "Aug 1971", "Nov 2021",   50.2,   98,     8.3,   "Y",
  86079, "MORNINGTON",                               -38.2358,  145.0667, "Jan 1962", "Aug 1975",   13.7,   84,     1.8,   "N",
  90176, "MORTLAKE RACECOURSE",                      -38.0737,  142.7744, "Jan 1990", "Nov 2021",   31.9,   99,    12.1,   "Y",
  85280, "MORWELL (LATROBE VALLEY AIRPORT)",         -38.2094,  146.4746, "Jan 1984", "Nov 2021",   37.9,   99,    10.6,   "Y",
  85291, "MOUNT BAW BAW",                            -37.8384,  146.2747, "Aug 1997", "Nov 2021",   23.2,   93,    12.9,   "Y",
  83023, "MOUNT BEAUTY",                             -36.7450,  147.1633, "Jan 1965", "Jan 1994",   28.4,   97,     2.0,   "N",
  83073, "MOUNT BUFFALO CHALET",                     -36.7220,  146.8189, "Jan 1965", "Apr 1980",   10.5,   65,     1.2,   "N",
  83024, "MOUNT BULLER",                             -37.1450,  146.4394, "Mar 1985", "Nov 2021",   35.7,   91,     9.7,   "Y",
  86243, "MOUNT DANDENONG GTV9",                     -37.8306,  145.3500, "Jan 1968", "Dec 1986",   19.0,   99,     1.8,   "N",
  86082, "MOUNT ELIZA",                              -38.2000,  145.1000, "Jan 1965", "Sep 1973",    8.8,   98,     1.0,   "N",
  83081, "MOUNT HOTHAM",                             -36.9833,  147.1478, "Dec 1977", "Jul 1990",   12.5,   88,     1.9,   "N",
  83085, "MOUNT HOTHAM",                             -36.9772,  147.1342, "Aug 1990", "Nov 2021",   30.3,   92,    11.7,   "Y",
  85296, "MOUNT MOORNAPA",                           -37.7481,  147.1428, "Dec 1993", "Nov 2021",   28.0,   98,    12.6,   "Y",
  84144, "MOUNT NOWA NOWA",                          -37.6924,  148.0908, "Nov 1995", "Nov 2021",   26.1,   99,    13.0,   "Y",
  82034, "MYRTLEFORD POST OFFICE",                   -36.5667,  146.7333, "Jan 1965", "Feb 1969",    4.2,   79,     1.9,   "N",
  88133, "NEWBURY",                                  -37.4167,  144.2833, "Jun 1971", "Aug 1974",    2.8,   80,     1.8,   "N",
  78031, "NHILL",                                    -36.3347,  141.6367, "Jun 1910", "Dec 2008",   97.6,   91,     3.3,   "N",
  78015, "NHILL AERODROME",                          -36.3093,  141.6486, "Jun 2003", "Nov 2021",   18.5,   99,    14.4,   "Y",
  85313, "NILMA NORTH (WARRAGUL)",                   -38.1321,  145.9865, "Jan 2014", "Nov 2021",    7.9,   99,     8.0,   "Y",
  85153, "NOOJEE",                                   -37.9000,  146.0000, "Jan 1965", "Sep 1975",   10.3,   95,     2.0,   "N",
  85277, "NOOJEE (SLIVAR)",                          -37.9039,  145.9719, "Oct 1981", "Apr 2012",   30.5,   98,     1.9,   "N",
  85269, "NOOJEE PEARCE",                            -37.8833,  146.0167, "Jul 1978", "Oct 1981",    3.3,   94,     1.5,   "N",
  84028, "NOWA NOWA",                                -37.7279,  148.0958, "Jan 1965", "Aug 1975",   10.7,   64,     1.6,   "N",
  80101, "NUMURKAH",                                 -36.0921,  145.4531, "Jul 1977", "Apr 1992",   14.8,   70,     1.0,   "N",
  80043, "NUMURKAH POST OFFICE",                     -36.1000,  145.4500, "Jan 1965", "Jun 1977",   12.5,   98,     1.0,   "N",
  85106, "OLSENS BRIDGE (MORWELL RIVER PRISON)",     -38.4861,  146.3233, "Jan 1962", "Apr 1997",   33.9,   94,     2.0,   "N",
  83090, "OMEO",                                     -37.1017,  147.6008, "Aug 2004", "Nov 2021",   17.3,  100,    14.9,   "Y",
  83025, "OMEO COMPARISON",                          -37.1011,  147.5980, "Jan 1957", "Dec 2009",   52.7,   92,     1.9,   "N",
  84145, "ORBOST",                                   -37.6922,  148.4667, "Nov 2000", "Nov 2021",   21.1,   99,    14.2,   "Y",
  84030, "ORBOST (COMPARISON)",                      -37.6917,  148.4589, "Jan 1957", "Nov 2011",   54.9,   98,     2.6,   "N",
  76047, "OUYEN (POST OFFICE)",                      -35.0681,  142.3125, "Jan 1957", "Nov 2021",   64.9,   98,     1.9,   "N",
  86354, "PHILLIP ISLAND PENGUIN RESERVE",           -38.5058,  145.1503, "Mar 1981", "Feb 2018",   35.7,   66,     1.8,   "N",
  84070, "POINT HICKS (LIGHTHOUSE)",                 -37.8017,  149.2747, "Jul 1966", "Dec 2019",   53.5,   98,     2.8,   "N",
  87166, "POINT WILSON",                             -38.0958,  144.5361, "Mar 1991", "Jun 2019",   28.1,   96,     9.5,   "Y",
  90175, "PORT FAIRY AWS",                           -38.3906,  142.2347, "Oct 1990", "Nov 2021",   31.0,   98,    12.1,   "Y",
  90171, "PORTLAND (CASHMORE AIRPORT)",              -38.3148,  141.4705, "Jun 1982", "Nov 2021",   39.5,   99,     9.8,   "Y",
  90162, "PORTLAND AERO",                            -38.3833,  141.6333, "Oct 1973", "Jun 1982",    8.8,   96,     2.0,   "N",
  90192, "PORTLAND NTC AWS",                         -38.3439,  141.6136, "Apr 2011", "Nov 2021",   10.7,   99,    20.4,   "Y",
  86097, "PORTSEA QUARANTINE STATION",               -38.3167,  144.6667, "Jan 1965", "Apr 1978",   12.6,   86,     1.8,   "N",
  85099, "POUND CREEK",                              -38.6297,  145.8107, "Apr 2007", "Nov 2021",   14.7,   99,    17.0,   "Y",
  86094, "POWELLTOWN DNRE",                          -37.8615,  145.7443, "Jan 1965", "Jul 1978",   13.6,   97,     2.0,   "N",
  88167, "PUCKAPUNYAL LYON HILL (DEFENCE)",          -36.9381,  145.0539, "Oct 2018", "Nov 2021",    3.2,   98,    24.0,   "Y",
  88166, "PUCKAPUNYAL WEST (DEFENCE)",               -37.0177,  144.8546, "Oct 2018", "Nov 2021",    3.2,   98,    23.9,   "Y",
  79101, "PYRENEES (BEN NEVIS)",                     -37.2281,  143.2006, "Dec 2007", "Nov 2021",   14.0,   99,    16.5,   "Y",
  87054, "QUEENSCLIFF",                              -38.2714,  144.6621, "Jan 1957", "Feb 1968",   10.9,   86,     1.9,   "N",
  77035, "RAINBOW POST OFFICE",                      -35.8975,  141.9950, "Jan 1965", "Jun 1970",    5.4,   78,     1.0,   "N",
  88051, "REDESDALE",                                -37.0194,  144.5203, "Jun 1993", "Nov 2021",   28.5,   99,    12.5,   "Y",
  90092, "RENNICK",                                  -37.8408,  140.9717, "Jan 1965", "Sep 1975",   10.8,   72,     1.1,   "N",
  86373, "RHYLL",                                    -38.4612,  145.3101, "Jun 1990", "Nov 2021",   31.5,   99,    12.1,   "Y",
  76125, "ROBINVALE CONSOLIDATED SCHOOL",            -34.6000,  142.7833, "Dec 1970", "Feb 1980",    8.8,   88,     2.0,   "N",
  80049, "ROCHESTER",                                -36.3612,  144.7110, "Jan 1965", "Jul 1975",   10.3,   97,     1.0,   "N",
  88068, "RUBICON SEC",                              -37.3389,  145.8547, "Jan 1957", "May 1993",   36.4,   99,     1.8,   "Y",
  82039, "RUTHERGLEN RESEARCH",                      -36.1048,  146.5094, "Jul 1939", "Nov 2021",   72.8,   74,     5.1,   "Y",
  86104, "SCORESBY RESEARCH INSTITUTE",              -37.8710,  145.2561, "Jan 1965", "Nov 2021",   54.6,   93,     6.6,   "Y",
  80099, "SERPENTINE LODDON VALLEY HWAY",           -36.4093,  143.9752, "Apr 1969", "Sep 1971",    2.5,   95,     2.0,   "N",
  78034, "SERVICETON",                               -36.3667,  140.9833, "Jan 1957", "Sep 1971",   14.8,  100,     2.0,   "N",
  88053, "SEYMOUR SHIRE DEPOT",                      -37.0277,  145.1491, "Jan 1965", "Sep 1970",    5.7,   79,     1.6,   "N",
  82092, "SHELLEY",                                  -36.1821,  147.5515, "Jan 1965", "Sep 1971",    5.2,   49,     1.7,   "N",
  87168, "SHEOAKS",                                  -37.9075,  144.1303, "Jun 1990", "Nov 2021",   31.5,   98,    12.0,   "Y",
  81125, "SHEPPARTON AIRPORT",                       -36.4288,  145.3949, "Jul 1996", "Nov 2021",   25.4,   99,    12.9,   "Y",
  86344, "SOUTH CHANNEL ISLAND",                     -38.3065,  144.8016, "May 1976", "Nov 2021",   38.6,   81,    15.6,   "Y",
  88153, "SPRING CREEK BASIN TWO",                   -37.0748,  145.7220, "Dec 1973", "Jun 1984",   10.6,   96,     1.9,   "N",
  79040, "ST ARNAUD",                                -36.6175,  143.2644, "Jan 1957", "Dec 1982",   25.9,   91,     1.6,   "N",
  86220, "ST KILDA HARBOUR - RMYS",                  -37.8640,  144.9639, "Mar 2006", "Nov 2021",   15.8,   99,    21.4,   "Y",
  82059, "STANLEY STATE FOREST",                     -36.4167,  146.7667, "Jan 1965", "Jun 1966",    1.5,   84,     1.0,   "N",
  79042, "STAWELL",                                  -37.1000,  142.8000, "Jan 1965", "May 1969",    4.4,   96,     1.6,   "N",
  79080, "STAWELL",                                  -37.0717,  142.7908, "Jun 1969", "Feb 1998",   28.7,   99,     2.0,   "N",
  79105, "STAWELL AERODROME",                        -37.0720,  142.7402, "Feb 1996", "Nov 2021",   25.8,   99,    11.0,   "Y",
  86306, "STONY POINT",                              -38.3833,  145.2167, "Aug 1972", "Feb 1983",   10.3,   96,     2.0,   "N",
  82042, "STRATHBOGIE",                              -36.8472,  145.7307, "Jan 1974", "Nov 2021",   47.9,   98,     1.5,   "N",
  77094, "SWAN HILL AERODROME",                      -35.3766,  143.5416, "Dec 1996", "Nov 2021",   25.0,   99,    12.3,   "Y",
  77042, "SWAN HILL POST OFFICE",                    -35.3406,  143.5533, "Jan 1960", "Dec 1996",   34.2,   90,     3.2,   "N",
  84087, "TABBERABBERA (THE PINES)",                 -37.5833,  147.3667, "Apr 1974", "Jul 1980",    5.8,   77,     1.9,   "N",
  85079, "TANJIL BREN",                              -37.8000,  146.2000, "Jan 1965", "Aug 1971",    5.2,   76,     1.0,   "N",
  86364, "TARRAWARRA MONASTERY",                     -37.6594,  145.4457, "Dec 1986", "Apr 1997",    8.9,   84,     1.6,   "N",
  81049, "TATURA INST SUSTAINABLE AG",               -36.4379,  145.2673, "Jan 1965", "Nov 2021",   56.9,   97,     5.4,   "Y",
  90077, "TERANG",                                   -38.2362,  142.9166, "Jan 1965", "Sep 1975",   10.8,   79,     1.0,   "N",
  85082, "TIDAL RIVER",                              -39.0310,  146.3227, "Jul 1965", "Jan 1972",    4.0,   45,     1.7,   "N",
  86142, "TOOLANGI (MOUNT ST LEONARD DPI)",          -37.5720,  145.5014, "Feb 1965", "Nov 2006",   41.8,   85,     1.0,   "N",
  86116, "TOORADIN",                                 -38.1667,  145.4000, "Jan 1957", "Oct 1974",   17.8,   87,     1.7,   "N",
  86068, "VIEWBANK",                                 -37.7408,  145.0972, "Nov 1999", "Nov 2021",   22.1,   99,    13.4,   "Y",
  78045, "WAIL",                                     -36.5167,  142.1667, 'Jan 1964', "Jun 1969",    4.6,   54,     1.0,   "N",
  88162, "WALLAN (KILMORE GAP)",                     -37.3807,  144.9654, 'Dec 1993', "Nov 2021",   28.0,   99,    10.8,   "Y",
  76064, "WALPEUP RESEARCH",                         -35.1201,  142.0040, 'Jan 1965', "Nov 2021",   56.9,   89,     6.1,   "Y",
  81115, "WANALTA DAEN STATION",                     -36.6284,  144.8727, 'Aug 1974', "Jun 1984",    9.9,  100,     2.0,   "N",
  82053, "WANGARATTA",                               -36.3667,  146.3000, 'Jan 1957', "May 1987",   30.1,   92,     1.9,   "N",
  82138, "WANGARATTA AERO",                          -36.4205,  146.3056, 'May 1987', "Nov 2021",   34.6,   99,     9.8,   "Y",
  78077, "WARRACKNABEAL MUSEUM",                     -36.2614,  142.4050, 'Jun 1969', "Oct 2019",   50.3,   99,     1.9,   "N",
  85093, "WARRAGUL",                                 -38.1731,  145.9487, 'Jan 1962', "Dec 1979",   16.0,   80,     1.9,   "N",
  89094, "WARRAMBINE BASIN NO 3",                    -37.8333,  143.8833, 'Mar 1972', "May 1982",   10.2,   99,     2.0,   "N",
  89092, "WARRAMBINE NO 2",                          -37.8783,  143.9023, 'Jun 1982', "Aug 1984",    2.3,  100,     2.0,   "N",
  90082, "WARRNAMBOOL (POST OFFICE)",                -38.3836,  142.4814, 'Jan 1957', "Apr 1983",   26.3,   99,     2.0,   "N",
  90172, "WARRNAMBOOL AIRPORT",                      -38.2919,  142.4353, 'Mar 1983', "Jun 2000",   17.3,  100,     7.0,   "N",
  90186, "WARRNAMBOOL AIRPORT NDB",                  -38.2869,  142.4525, 'Oct 1998', "Nov 2021",   23.2,   99,    13.1,   "Y",
  86122, "WATSONIA LOYOLA",                          -37.7000,  145.0833, 'Jan 1965', "Dec 1973",    9.0,   97,     1.0,   "N",
  90083, "WEEAPROINAH",                              -38.6357,  143.5127, 'Nov 1981', "Feb 2012",   30.3,   99,     1.5,   "N",
  87065, "WERRIBEE RESEARCH FARM",                   -37.9000,  144.6833, 'Jan 1965', "Jul 1980",   15.6,   98,     1.0,   "N",
  89112, "WESTMERE",                                 -37.7067,  142.9378, 'Feb 2006', "Nov 2021",   15.8,   99,    16.4,   "Y",
  85283, "WILLOW GROVE (BLUE ROCK RESERVOIR)",       -38.0809,  146.2139, 'Oct 1985', "Jul 2019",   32.0,   70,     1.0,   "N",
  85096, "WILSONS PROMONTORY LIGHTHOUSE",            -39.1297,  146.4246, 'Jan 1957', "Nov 2021",   64.7,   97,     4.9,   "Y",
  82056, "WODONGA",                                  -36.1180,  146.9061, 'Jan 1957', "May 1968",   11.4,   98,     2.0,   "N",
  85152, "WON WRON PRISON",                          -38.4856,  146.6717, 'Feb 1967', "Sep 1975",    8.6,   98,     2.0,   "N",
  86127, "WONTHAGGI",                                -38.6078,  145.5965, 'Jul 1968', "Nov 2021",   53.3,   96,     3.9,   "N",
  83033, "WOODS POINT",                              -37.5695,  146.2543, 'Jan 1957', "Sep 1969",   12.8,   95,     2.0,   "N",
  85097, "WOODSIDE",                                 -38.5000,  146.9000, 'Jun 1924', "Dec 1936",    2.2,   12,     1.0,   "N",
  89036, "WOOHLPOOER",                               -37.3333,  142.1500, 'Jan 1965', "Aug 1975",    9.9,   71,     1.0,   "N",
  85098, "YALLOURN",                                 -38.2000,  146.4000, 'Apr 1932', "Oct 1949",   10.8,   54,     1.8,   "N",
  85103, "YALLOURN SEC",                             -38.1858,  146.3317, 'Jan 1957', "Jun 1986",   29.4,   97,     2.0,   "N",
  85151, "YARRAM AIRPORT",                           -38.5647,  146.7479, 'Oct 2007', "Nov 2021",   14.2,   99,    16.4,   "Y",
  81124, "YARRAWONGA",                               -36.0294,  146.0305, 'May 1993', "Nov 2021",   28.4,   97,    10.7,   "Y",
  81057, "YARRAWONGA POST OFFICE",                   -36.0281,  146.0044, 'Feb 1965', "Sep 1975",   10.7,   86,     1.7,   "N"
)


metar_raw <- tribble(

  ~Site,    ~Name,                                     ~Lat,       ~Lon,      ~Start,    ~End,       ~Years,   ~percent,      ~Obs, ~AWS,

  90180, "AIREYS INLET",                             -38.4583,  144.0883, "Sep 1991", "Nov 2021",   30.3,   99,    50.3,   "Y",
  88001, "ALEXANDRA (POST OFFICE)",                  -37.1916,  145.7116, "Jun 1999", "Nov 1999",    0.5,   13,     1.5,   "N",
  87113, "AVALON AIRPORT",                           -38.0288,  144.4783, "Oct 1994", "Nov 2021",   27.2,   99,    46.9,   "Y",
  85279, "BAIRNSDALE AIRPORT",                       -37.8818,  147.5669, "May 1997", "Nov 2021",   24.2,   97,    49.9,   "Y",
  89002, "BALLARAT AERODROME",                       -37.5127,  143.7911, "Jul 2000", "Nov 2021",   21.4,   99,    59.2,   "Y",
  81123, "BENDIGO AIRPORT",                          -36.7411,  144.3274, "Nov 1993", "Nov 2021",   28.1,   98,    44.9,   "Y",
  87184, "BREAKWATER (GEELONG RACECOURSE)",          -38.1737,  144.3765, "Jun 2011", "Nov 2021",   10.5,  100,    49.2,   "Y",
  90184, "CAPE NELSON LIGHTHOUSE",                   -38.4306,  141.5437, "May 1997", "Nov 2021",   24.6,   99,    52.6,   "Y",
  90015, "CAPE OTWAY LIGHTHOUSE",                    -38.8557,  143.5130, "Aug 1994", "Nov 2021",   27.3,   99,    51.3,   "Y",
  90182, "CASTERTON",                                -37.5831,  141.3340, "Mar 2005", "Nov 2021",   16.8,   99,    49.3,   "Y",
  86361, "CERBERUS",                                 -38.3646,  145.1785, "Sep 1991", "Nov 2021",   30.2,   98,    44.7,   "Y",
  80128, "CHARLTON",                                 -36.2846,  143.3341, "Jul 2004", "Nov 2021",   17.4,   99,    48.7,   "Y",
  90035, "COLAC (MOUNT GELLIBRAND)",                 -38.2332,  143.7924, "Jul 2000", "Nov 2021",   21.4,   99,    49.9,   "Y",
  86383, "COLDSTREAM",                               -37.7239,  145.4092, "Oct 1994", "Nov 2021",   27.2,   99,    44.0,   "Y",
  84143, "COMBIENBAR AWS",                           -37.3416,  149.0227, "Jul 1994", "Nov 2021",   27.3,   98,    43.1,   "Y",
  85301, "CORNER INLET (YANAKIE)",                   -38.8051,  146.1936, "Jan 2013", "Nov 2021",    8.9,   99,    50.5,   "Y",
  90194, "DARTMOOR",                                 -37.9222,  141.2614, "Mar 2009", "Nov 2021",   12.8,   99,    48.9,   "Y",
  78072, "DONALD",                                   -36.3611,  143.0038, "Feb 1999", "Feb 2001",    2.1,   94,    48.7,   "N",
  85072, "EAST SALE",                                -38.1156,  147.1323, "Nov 1993", "Sep 2017",   23.9,   98,    48.7,   "Y",
  85314, "EAST SALE AIRPORT",                        -38.1016,  147.1398, "Oct 2017", "Nov 2021",    4.2,  100,    53.5,   "Y",
  79099, "EDENHOPE AIRPORT",                         -37.0222,  141.2657, "Apr 2005", "Nov 2021",   16.7,   99,    48.8,   "Y",
  88164, "EILDON FIRE TOWER",                        -37.2091,  145.8423, "Dec 1995", "Nov 2021",   26.0,   99,    45.7,   "Y",
  86038, "ESSENDON AIRPORT",                         -37.7276,  144.9066, "Feb 1999", "Nov 2021",   20.0,   83,    49.5,   "Y",
  83084, "FALLS CREEK",                              -36.8708,  147.2755, "Sep 1991", "Nov 2021",   29.7,   94,    50.6,   "Y",
  86376, "FAWKNER BEACON",                           -37.9483,  144.9269, "Mar 1992", "Nov 2021",   28.7,   93,    47.4,   "Y",
  86266, "FERNY CREEK",                              -37.8748,  145.3497, "Mar 2011", "Nov 2021",   10.8,   99,    48.0,   "Y",
  86372, "FERNY CREEK (DUNNS HILL)",                 -37.8775,  145.3364, "Sep 1991", "Jul 2011",   19.9,   98,    50.1,   "Y",
  86388, "FRANKSTON (BALLAM PARK)",                  -38.1517,  145.1615, "Jul 2020", "Nov 2021",    1.4,   99,    48.1,   "Y",
  86371, "FRANKSTON BEACH",                          -38.1480,  145.1152, "Sep 1991", "Nov 2021",   30.3,   98,    48.3,   "Y",
  84016, "GABO ISLAND LIGHTHOUSE",                   -37.5679,  149.9158, "Aug 2007", "Nov 2021",   14.3,   99,    50.5,   "Y",
  84142, "GELANTIPY",                                -37.2200,  148.2626, "May 1992", "Nov 2021",   29.5,   98,    42.8,   "Y",
  79103, "GRAMPIANS (MOUNT WILLIAM)",                -37.2950,  142.6039, "Dec 2005", "Nov 2021",   15.8,   97,    53.6,   "Y",
  87163, "GROVEDALE (GEELONG AIRPORT)",              -38.2241,  144.3345, "Sep 1991", "Aug 2011",   20.0,   97,    48.1,   "Y",
  90173, "HAMILTON AIRPORT",                         -37.6486,  142.0636, "May 1995", "Nov 2021",   26.6,   99,    48.8,   "Y",
  77010, "HOPETOUN AIRPORT",                         -35.7151,  142.3569, "Jun 2004", "Nov 2021",   17.5,   99,    49.2,   "Y",
  79100, "HORSHAM AERODROME",                        -36.6699,  142.1733, "Oct 1997", "Nov 2021",   24.2,   99,    47.2,   "Y",
  82139, "HUNTERS HILL",                             -36.2137,  147.5395, "Apr 1993", "Nov 2021",   28.7,   98,    42.1,   "Y",
  79097, "KANAGULK",                                 -37.1169,  141.8031, "May 2004", "Nov 2021",   17.6,   99,    48.7,   "Y",
  78018, "KELLALAC (WARRACKNABEAL AIRPORT)",         -36.3204,  142.4161, "Dec 2016", "Nov 2021",    5.0,   99,    49.7,   "Y",
  80091, "KYABRAM",                                  -36.3350,  145.0638, "Sep 2002", "Nov 2021",   16.0,   78,    45.3,   "Y",
  87031, "LAVERTON RAAF",                            -37.8565,  144.7565, "Nov 1993", "Nov 2021",   28.1,   99,    46.0,   "Y",
  79028, "LONGERENONG",                              -36.6722,  142.2991, "May 1997", "Nov 2021",   24.6,  100,    48.7,   "Y",
  89105, "LOOKOUT HILL",                             -37.2822,  143.2469, "Feb 1992", "May 2007",   15.0,   94,    50.2,   "Y",
  84084, "MALLACOOTA",                               -37.5976,  149.7289, "Nov 1993", "Nov 2021",   28.1,   98,    43.6,   "Y",
  88109, "MANGALORE AIRPORT",                        -36.8886,  145.1859, "Nov 1993", "Nov 2021",   28.1,   99,    49.3,   "Y",
  86338, "MELBOURNE (OLYMPIC PARK)",                 -37.8255,  144.9816, "May 2013", "Nov 2021",    8.6,   99,    49.1,   "Y",
  86282, "MELBOURNE AIRPORT",                        -37.6654,  144.8322, "Jul 1970", "Nov 2021",   51.4,  100,    48.7,   "Y",
  86397, "MELBOURNE AIRPORT (EAST ANEMOMETER)",      -37.6602,  144.8443, "Aug 2009", "Nov 2021",    7.6,   61,    50.6,   "Y",
  86395, "MELBOURNE AIRPORT (NORTH ANEMOMETER)",     -37.6577,  144.8333, "Aug 2009", "Nov 2021",    7.6,   60,    50.1,   "Y",
  86396, "MELBOURNE AIRPORT (SOUTH ANEMOMETER)",     -37.6832,  144.8383, "Aug 2009", "Nov 2021",    7.6,   61,    49.9,   "Y",
  86398, "MELBOURNE AIRPORT (WEST ANEMOMETER)",      -37.6624,  144.8193, "Aug 2009", "Nov 2021",    7.6,   60,    50.6,   "Y",
  86381, "MELBOURNE HARBOUR CONTROL",                -37.8208,  144.9325, "Mar 1992", "Nov 1999",    6.5,   73,    46.3,   "Y",
  86071, "MELBOURNE REGIONAL OFFICE",                -37.8075,  144.9700, "Mar 1997", "Apr 2012",   11.4,   71,    46.3,   "Y",
  76031, "MILDURA AIRPORT",                          -34.2358,  142.0867, "Jan 1989", "Nov 2021",   32.3,   97,    42.7,   "Y",
  86077, "MOORABBIN AIRPORT",                        -37.9800,  145.0962, "Aug 1992", "Nov 2021",   29.3,   99,    52.0,   "Y",
  90176, "MORTLAKE RACECOURSE",                      -38.0737,  142.7744, "Sep 1991", "Nov 2021",   30.3,   98,    48.4,   "Y",
  85280, "MORWELL (LATROBE VALLEY AIRPORT)",         -38.2094,  146.4746, "May 1997", "Nov 2021",   24.6,   99,    49.3,   "Y",
  85291, "MOUNT BAW BAW",                            -37.8384,  146.2747, "Jan 1994", "Nov 2021",   26.2,   88,    44.0,   "Y",
  83024, "MOUNT BULLER",                             -37.1450,  146.4394, "Oct 1992", "Nov 2021",   28.4,   91,    46.4,   "Y",
  83085, "MOUNT HOTHAM",                             -36.9772,  147.1342, "Sep 1991", "Nov 2021",   29.5,   94,    45.9,   "Y",
  85296, "MOUNT MOORNAPA",                           -37.7481,  147.1428, "Jan 1994", "Nov 2021",   27.9,   99,    42.5,   "Y",
  84144, "MOUNT NOWA NOWA",                          -37.6924,  148.0908, "Oct 1995", "Nov 2021",   26.2,   99,    44.0,   "Y",
  78015, "NHILL AERODROME",                          -36.3093,  141.6486, "Jun 2003", "Nov 2021",   18.5,   99,    49.3,   "Y",
  85313, "NILMA NORTH (WARRAGUL)",                   -38.1321,  145.9865, "Jan 2014", "Nov 2021",    7.9,   99,    49.0,   "Y",
  83090, "OMEO",                                     -37.1017,  147.6008, "Aug 2004", "Nov 2021",   17.3,   99,    49.0,   "Y",
  84145, "ORBOST",                                   -37.6922,  148.4667, "Nov 2000", "Nov 2021",   21.1,   99,    48.3,   "Y",
  87166, "POINT WILSON",                             -38.0958,  144.5361, "Sep 1991", "Jun 2019",   27.8,   97,    47.5,   "Y",
  90175, "PORT FAIRY AWS",                           -38.3906,  142.2347, "Sep 1991", "Nov 2021",   30.1,   98,    49.2,   "Y",
  90171, "PORTLAND (CASHMORE AIRPORT)",              -38.3148,  141.4705, "Oct 1995", "Nov 2021",   26.2,   99,    53.0,   "Y",
  90192, "PORTLAND NTC AWS",                         -38.3439,  141.6136, "Apr 2011", "Nov 2021",   10.7,   99,    50.7,   "Y",
  85099, "POUND CREEK",                              -38.6297,  145.8107, "Apr 2007", "Nov 2021",   14.7,   99,    49.5,   "Y",
  88167, "PUCKAPUNYAL LYON HILL (DEFENCE)",          -36.9381,  145.0539, "Oct 2018", "Nov 2021",    3.2,   98,    49.6,   "Y",
  88166, "PUCKAPUNYAL WEST (DEFENCE)",               -37.0177,  144.8546, "Oct 2018", "Nov 2021",    3.2,   98,    50.3,   "Y",
  79101, "PYRENEES (BEN NEVIS)",                     -37.2281,  143.2006, "Dec 2007", "Nov 2021",   14.0,   99,    51.0,   "Y",
  88051, "REDESDALE",                                -37.0194,  144.5203, "Nov 1994", "Nov 2021",   23.0,   84,    45.7,   "Y",
  86373, "RHYLL",                                    -38.4612,  145.3101, "Sep 1991", "Nov 2021",   30.3,   98,    48.1,   "Y",
  82039, "RUTHERGLEN RESEARCH",                      -36.1048,  146.5094, "Jan 1998", "Nov 2021",   23.8,   98,    44.5,   "Y",
  86104, "SCORESBY RESEARCH INSTITUTE",              -37.8710,  145.2561, "Sep 1996", "Nov 2021",   25.3,   99,    48.0,   "Y",
  87168, "SHEOAKS",                                  -37.9075,  144.1303, "Sep 1991", "Nov 2021",   30.0,   96,    48.6,   "Y",
  81125, "SHEPPARTON AIRPORT",                       -36.4288,  145.3949, "Jul 1996", "Nov 2021",   25.4,   99,    45.3,   "Y",
  86344, "SOUTH CHANNEL ISLAND",                     -38.3065,  144.8016, "Dec 1998", "Nov 2021",   22.9,   98,    55.7,   "Y",
  86220, "ST KILDA HARBOUR - RMYS",                  -37.8640,  144.9639, "Mar 2006", "Nov 2021",   15.8,   99,    50.6,   "Y",
  79105, "STAWELL AERODROME",                        -37.0720,  142.7402, "Feb 1996", "Nov 2021",   25.8,   99,    48.0,   "Y",
  77094, "SWAN HILL AERODROME",                      -35.3766,  143.5416, "May 1997", "Nov 2021",   24.6,   99,    45.7,   "Y",
  81049, "TATURA INST SUSTAINABLE AG",               -36.4379,  145.2673, "May 2006", "Nov 2021",   11.8,   72,    47.5,   "Y",
  86068, "VIEWBANK",                                 -37.7408,  145.0972, "Jul 2000", "Nov 2021",   21.4,   99,    47.8,   "Y",
  88162, "WALLAN (KILMORE GAP)",                     -37.3807,  144.9654, "Nov 1994", "Nov 2021",   25.7,   93,    51.6,   "Y",
  76064, "WALPEUP RESEARCH",                         -35.1201,  142.0040, "Nov 1998", "Nov 2021",   23.1,   99,    46.0,   "Y",
  82138, "WANGARATTA AERO",                          -36.4205,  146.3056, "Oct 1995", "Nov 2021",   26.2,   99,    45.0,   "Y",
  90186, "WARRNAMBOOL AIRPORT NDB",                  -38.2869,  142.4525, "Sep 1998", "Nov 2021",   23.3,  100,    51.7,   "Y",
  89112, "WESTMERE",                                 -37.7067,  142.9378, "Feb 2006", "Nov 2021",   15.8,   99,    49.0,   "Y",
  85096, "WILSONS PROMONTORY LIGHTHOUSE",            -39.1297,  146.4246, "Sep 2000", "Nov 2021",   21.0,   97,    55.9,   "Y",
  78087, "YANAC NORTH COMPARISON",                   -36.1147,  141.4612, "May 1997", "Jun 1999",    1.8,   72,    45.3,   "Y",
  85151, "YARRAM AIRPORT",                           -38.5647,  146.7479, "Oct 2007", "Nov 2021",   14.2,   99,    50.5,   "Y",
  81124, "YARRAWONGA",                               -36.0294,  146.0305, "Jul 2002", "Nov 2021",   19.4,   99,    48.5,   "Y"
)

wind <- wind_raw %>%
  mutate(Start = tsibble::yearmonth(Start),
         End = tsibble::yearmonth(End)) %>%
  filter(lubridate::year(Start) < 2019,
         lubridate::year(End) >= 2020)


metar <- metar_raw %>%
  mutate(Start = tsibble::yearmonth(Start),
         End = tsibble::yearmonth(End)) %>%
  filter(lubridate::year(Start) < 2019,
         lubridate::year(End) >= 2020)

library(cubble)
vic_map <- ozmaps::abs_ste %>% filter(NAME == "Victoria")
wind_meta <- weatherdata::wind %>%
  as_tibble() %>%
  select(-ts)

wind_data <- weatherdata::wind %>%
  stretch() %>%
  filter(between(lubridate::date(time), as.Date("2019-10-01"), as.Date("2020-03-31")),
         direction <= 360) %>%
  as_tibble()

plot_map(vic_map) +
  #geom_point(data = wind_meta, aes(x = lon, y = lat), color = "red") +
  geom_point(data = metar, aes(x = Lon, y = Lat), color = "red") +
  geom_point(data = wind_meta, aes(x = lon, y = lat), size = 0.8)
